pyproject.toml
[project]
name = "gm-chatbot"
version = "2.0.0"
description = "Game Master Assistant for tabletop RPGs"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0",
    "openai>=1.50.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "typer>=0.13.0",
    "structlog>=24.0.0",
    "scalar-fastapi>=1.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "ty>=0.2.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "ty>=0.2.0",
    "pre-commit>=4.0.0",
]

[tool.ty]
python-version = "3.12"
strict = true
warn-unreachable = true
